refactor: 파일 위치 변경 및 누락된 파일 이름 수정 추가
