+++ data/src/main/java/com/housweet/data/datasource/AppSettingRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.housweet.data.network
+package com.housweet.data.datasource
 
 import com.housweet.data.response.AppSettingItemResponse
 import com.housweet.data.response.AppSettingResponse
