+++ data/src/main/java/com/housweet/data/datasource/AuthRemoteDataSourceImpl.kt
@@ -1,6 +1,7 @@
-package com.housweet.data.network
+package com.housweet.data.datasource
 
 import com.housweet.data.BuildConfig
+import com.housweet.data.network.KtorService
 import com.housweet.data.request.AgreeTermsOfServiceRequest
 import com.housweet.data.response.IsTermsOfServiceAgreedResponse
 import com.housweet.data.request.KakaoLoginRequest
