+++ data/src/main/java/com/housweet/data/datasource/ChatRemoteDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.housweet.data.network
+package com.housweet.data.datasource
 
 import android.util.Log
 import com.housweet.data.response.ChatUserResponse
