+++ data/src/main/java/com/housweet/data/datasource/HouseRegisterRemoteDataSource.kt
@@ -1,10 +1,11 @@
-package com.housweet.data.network
+package com.housweet.data.datasource
 
 import com.housweet.data.request.RegisterHouseRequest
 import com.housweet.data.request.UpdateHouseRequest
+import com.housweet.data.response.PostingDetailResponse
 
 interface HouseRegisterRemoteDataSource {
     suspend fun registerHouse(body: RegisterHouseRequest)
-    suspend fun getPostingDetail(id: Int): PostingDetailDto
+    suspend fun getPostingDetail(id: Int): PostingDetailResponse
     suspend fun updateHouse(id: Int, body: UpdateHouseRequest)
 }
