+++ data/src/main/java/com/housweet/data/datasource/ImageUploadRemoteDataSourceImpl.kt
@@ -2,12 +2,9 @@ package com.housweet.data.datasource
 
 import android.util.Log
 import com.housweet.data.api.ImageUploadApiService
-import com.housweet.data.network.ImageUploadRemoteDataSource
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
-import okhttp3.RequestBody
 import okhttp3.RequestBody.Companion.asRequestBody
-import okhttp3.internal.http.hasBody
 import javax.inject.Inject
 import java.io.File
 import java.net.URLConnection
