+++ data/src/main/java/com/housweet/data/datasource/NoticeRemoteDataSourceImpl.kt
@@ -1,6 +1,7 @@
-package com.housweet.data.network
+package com.housweet.data.datasource
 
 import com.housweet.data.BuildConfig
+import com.housweet.data.network.KtorService
 import com.housweet.data.response.NoticeResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
