+++ data/src/main/java/com/housweet/data/datasource/RoomRemoteDataSourceImpl.kt
@@ -1,5 +1,6 @@
-package com.housweet.data.network
+package com.housweet.data.datasource
 
+import com.housweet.data.network.KtorService
 import com.housweet.data.response.MyRoomResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
