+++ data/src/main/java/com/housweet/data/repository/AuthRepositoryImpl.kt
@@ -2,8 +2,7 @@ package com.housweet.data.repository
 
 import android.util.Log
 import com.housweet.data.local.AuthLocalDataSource
-import com.housweet.data.network.AuthRemoteDataSource
-import com.housweet.data.network.RoomRepository
+import com.housweet.data.datasource.AuthRemoteDataSource
 import com.housweet.data.response.LoginResponse
 import com.housweet.data.response.toAuthToken
 import com.housweet.data.utils.NetworkConnectionManager
