+++ data/src/main/java/com/housweet/data/repository/HouseRegisterRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.housweet.data.repository
 
 import com.housweet.data.request.RegisterHouseRequest
-import com.housweet.data.network.HouseRegisterRemoteDataSource
+import com.housweet.data.datasource.HouseRegisterRemoteDataSource
 import com.housweet.data.request.UpdateHouseRequest
 import com.housweet.data.response.toDomain
 import com.housweet.domain.model.HouseRegisterModel
