+++ data/src/main/java/com/housweet/data/repository/ImageUploadRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.housweet.data.repository
 
-import com.housweet.data.network.ImageUploadRemoteDataSource
+import com.housweet.data.datasource.ImageUploadRemoteDataSource
 import com.housweet.domain.repository.ImageUploadRepository
 import java.io.File
 import javax.inject.Inject
