+++ data/src/main/java/com/housweet/data/repository/NoticeRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.housweet.data.repository
 
-import com.housweet.data.network.NoticeRemoteDataSource
+import com.housweet.data.datasource.NoticeRemoteDataSource
 import com.housweet.data.response.toDomain
 import com.housweet.domain.model.Notice
 import com.housweet.domain.repository.NoticeRepository
