+++ data/src/main/java/com/housweet/data/repository/ReportRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.housweet.data.repository
 
-import com.housweet.data.network.ReportRemoteDataSource
+import com.housweet.data.datasource.ReportRemoteDataSource
 import com.housweet.data.request.ReportRequest
 import com.housweet.domain.model.Report
 import com.housweet.domain.repository.ReportRepository
